Show IPy widget warning only when loading ext
